==> notebooks/8. metacell_plot.ipynb <==
import pandas as pd
import scanpy as sc
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns

import pathlib as pl

from statannotations.Annotator import Annotator

def pretty_ax(ax):
    ax.spines['right'].set_visible(False)
    ax.spines['top'].set_visible(False)
    ax.tick_params(
        axis='both',  
        which='both',      
        bottom=True,     
        top=False,
        left=False,
        labelbottom=True,
        labelleft = True)
    ax.spines["bottom"].set_linewidth(1.5)
    ax.spines["left"].set_linewidth(1.5)

resdir = pl.Path("/add/path/here/info-metacells")

meta_df = []
mapping = {"breast": "Breast", "metpancreas": "Met. Pancreas", 
           "pancreas": "Pancreas", "rcc": "RCC", "sclc": "SCLC", 
           "uvealmelanoma": "Uveal Melanoma", "luad": "LUAD"}
for f in resdir.iterdir():
    if f.stem!=".DS_Store":
        df = pd.read_csv(f, index_col=0)
        df["Cancer"] = mapping[f.stem]
        meta_df.append(df)

meta_df = pd.concat(meta_df)

meta_df["HighMT"] = meta_df["HighMT"].replace({0: "LowMT", 1: "HighMT"})
meta_df["Malignant"] = meta_df["Malignant"].replace({0: "TME", 1: "Malignant"})

fig, ax = plt.subplots(1,1,figsize=(5,2.75))

pairs = []
for ct in meta_df.Cancer.unique():
    pairs.append(((ct,"HighMT"),(ct,"LowMT")))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="Dissociation stress", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="Dissociation stress", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
fig.tight_layout()
fig.savefig("/add/path/here/figures/dissociation_stress_allstudies.svg", dpi=300, bbox_inches="tight")

mal_df = meta_df[meta_df.Malignant=="Malignant"]

corrs = {}
for ct in ["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"]:
    r = mal_df[mal_df["Cancer"]==ct][["Dissociation stress",
                                      "HighMT"]].replace({"LowMT": 0, "HighMT": 1}).corr().loc["HighMT","Dissociation stress"]
    corrs[ct] = [r]

pd.DataFrame(corrs)

meta_df.columns

fig, ax = plt.subplots(1,1,figsize=(5,2.75))

pairs = []
for ct in meta_df.Cancer.unique():
    pairs.append(((ct,"HighMT"),(ct,"LowMT")))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="Dissociation stress", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="Dissociation stress", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
fig.tight_layout()
fig.savefig("/add/path/here/figures/dissociation_stress_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))

pairs = []
for ct in meta_df.Cancer.unique():
    pairs.append(((ct,"HighMT"),(ct,"LowMT")))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="Mito transfer", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","LUAD","SCLC","RCC","Breast","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="Mito transfer", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
fig.tight_layout()
fig.savefig("/add/path/here/figures/mito_transfer_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))

pairs = []
for ct in meta_df.Cancer.unique():
    pairs.append(((ct,"HighMT"),(ct,"LowMT")))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="Fission", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","LUAD","SCLC","RCC","Breast","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="Fission", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
fig.tight_layout()
fig.savefig("/add/path/here/figures/fission_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="O-Flanagan_dissociation_stress", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="O-Flanagan_dissociation_stress", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
fig.tight_layout()
fig.savefig("/add/path/here/figures/oflanagan_dissociation_stress_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="Machado_dissociation_stress", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="Machado_dissociation_stress", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
fig.tight_layout()
fig.savefig("/add/path/here/figures/machado_dissociation_stress_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="Van-den-Brink_dissociation_stress", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="Van-den-Brink_dissociation_stress", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
fig.tight_layout()
fig.savefig("/add/path/here/figures/vdbrink_dissociation_stress_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="CYP genes", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="CYP genes", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
ax.set_ylabel("CYP genes (phase I)")
fig.tight_layout()
fig.savefig("/add/path/here/figures/cypgenes_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="GST genes", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="GST genes", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
ax.set_ylabel("GST genes (phase II)")
fig.tight_layout()
fig.savefig("/add/path/here/figures/gstgenes_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="UGT genes", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="UGT genes", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
ax.set_ylabel("UGT genes (phase II)")
fig.tight_layout()
fig.savefig("/add/path/here/figures/ugtgenes_allstudies.svg", dpi=300, bbox_inches="tight")

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="ABC transporters", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="ABC transporters", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
ax.set_ylabel("ABC transporters (phase III)")
fig.tight_layout()
fig.savefig("/add/path/here/figures/abctransporters_allstudies.svg", dpi=300, bbox_inches="tight")

# EGFR expression

resdir = pl.Path("/add/path/here/egfr_expression_metacells/")

meta_df = []
mapping = {"Breast": "Breast", "MetPancreas": "Met. Pancreas", 
           "Pancreas": "Pancreas", "RCC": "RCC", "SCLC": "SCLC", 
           "UvealMelanoma": "Uveal Melanoma", "LUAD": "LUAD"}

for f in resdir.iterdir():
    if f.stem==".DS_Store":
        continue
    df = pd.read_csv(f, index_col=0).rename(columns={"ERBIN": "ERBB2IP"})
    df["Cancer"] = mapping[f.stem]
    meta_df.append(df)

meta_df = pd.concat(meta_df)

meta_df["HighMT"] = meta_df["HighMT"].replace({0: "LowMT", 1: "HighMT"})
meta_df["Malignant"] = meta_df["Malignant"].replace({0: "TME", 1: "Malignant"})

pairs = []
for ct in meta_df.Cancer.unique():
    pairs.append(((ct,"HighMT"),(ct,"LowMT")))

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="EGFR", hue_order=["LowMT","HighMT"], 
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="EGFR", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
ax.set_ylabel("EGFR")
fig.tight_layout()
fig.savefig("/add/path/here/figures/egfr_allstudies.svg", dpi=300, bbox_inches="tight")

pairs = []
for ct in meta_df.Cancer.unique():
    pairs.append(((ct,"HighMT"),(ct,"LowMT")))

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="ERBB2", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="ERBB2", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1.3,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
ax.set_ylabel("ERBB2")
fig.tight_layout()
fig.savefig("/add/path/here/figures/erbb2_allstudies.svg", dpi=300, bbox_inches="tight")

pairs = []
for ct in meta_df.Cancer.unique():
    pairs.append(((ct,"HighMT"),(ct,"LowMT")))

fig, ax = plt.subplots(1,1,figsize=(5,2.75))
sns.boxplot(data=meta_df[meta_df.Malignant=="Malignant"], 
            x="Cancer", y="ERBB3", hue_order=["LowMT","HighMT"],
            order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
            hue="HighMT", ax=ax)
pretty_ax(ax)
annot = Annotator(
        ax,
        pairs=pairs,
        data=meta_df[meta_df.Malignant=="Malignant"], 
        x="Cancer", y="ERBB3", hue_order=["LowMT","HighMT"],
        order=["Uveal Melanoma","SCLC","RCC","Breast","LUAD","Pancreas","Met. Pancreas"],
        hue="HighMT"
    )
annot.configure(
    test="Mann-Whitney",
    loc="inside",
    text_format="star",
    show_test_name=False,
    verbose=2,
    comparisons_correction=None,
    fontsize=10,
)
annot.apply_test()
_, test_results = annot.annotate()
ax.legend(frameon=False, bbox_to_anchor=(1,1.1,0,0))
ax.set_xticks(ax.get_xticks(), ax.get_xticklabels(), rotation=45, ha='right')
ax.set_xlabel("")
ax.set_ylabel("ERBB3")
fig.tight_layout()
fig.savefig("/add/path/here/figures/erbb3_allstudies.svg", dpi=300, bbox_inches="tight")
