==> notebooks/13.cell_lineage_analysis.ipynb <==
import scanpy as sc
import pandas as pd
import re
import numpy as np
import anndata as ad
from scipy.sparse import csr_matrix
from statannotations.Annotator import Annotator
import seaborn as sns
import matplotlib.pyplot as plt

# Kuramochi GSE223003

path = "/add/path/here/GSE223003_lineage/"

carbo1_pre_info =pd.read_csv(path+ "GSM6938163_ReSisTrace_Carboplatin_1_pre_treatment_cell_info.tsv.gz", sep='\t', index_col = 0)
carbo1_pre_umi = pd.read_csv(path+ "GSM6938163_ReSisTrace_Carboplatin_1_pre_treatment_UMIcounts.tsv.gz", sep='\t', index_col = 0)
carbo1_post_info = pd.read_csv(path+ "GSM6938164_ReSisTrace_Carboplatin_1_post_treatment_cell_info.tsv.gz", sep='\t', index_col = 0)
carbo1_post_umi = pd.read_csv(path+ "GSM6938164_ReSisTrace_Carboplatin_1_post_treatment_UMIcounts.tsv.gz", sep='\t', index_col=0)

carbo2_pre_info =pd.read_csv(path+ "GSM6938165_ReSisTrace_Carboplatin_2_pre_treatment_cell_info.tsv.gz", sep='\t', index_col = 0)
carbo2_pre_umi = pd.read_csv(path+ "GSM6938165_ReSisTrace_Carboplatin_2_pre_treatment_UMIcounts.tsv.gz", sep='\t', index_col=0)
carbo2_post_info = pd.read_csv(path+ "GSM6938166_ReSisTrace_Carboplatin_2_post_treatment_cell_info.tsv.gz", sep='\t', index_col = 0)
carbo2_post_umi = pd.read_csv(path+ "GSM6938166_ReSisTrace_Carboplatin_2_post_treatment_UMIcounts.tsv.gz", sep='\t', index_col=0)

ola1_pre_info = pd.read_csv(path+"GSM6938167_ReSisTrace_Olaparib_1_pre_treatment_cell_info.tsv.gz", sep='\t', index_col = 0)
ola1_pre_umi = pd.read_csv(path+"GSM6938167_ReSisTrace_Olaparib_1_pre_treatment_UMIcounts.tsv.gz", sep='\t', index_col=0)
ola1_post_info = pd.read_csv(path+"GSM6938168_ReSisTrace_Olaparib_1_post_treatment_cell_info.tsv.gz", sep='\t', index_col = 0)
ola1_post_umi = pd.read_csv(path+"GSM6938168_ReSisTrace_Olaparib_1_post_treatment_UMIcounts.tsv.gz", sep='\t', index_col=0)
ola2_pre_info = pd.read_csv(path+"GSM6938169_ReSisTrace_Olaparib_2_pre_treatment_cell_info.tsv.gz", sep='\t', index_col = 0)
ola2_pre_umi = pd.read_csv(path+"GSM6938169_ReSisTrace_Olaparib_2_pre_treatment_UMIcounts.tsv.gz", sep='\t', index_col=0)
ola2_post_info = pd.read_csv(path+"GSM6938170_ReSisTrace_Olaparib_2_post_treatment_cell_info.tsv.gz", sep='\t', index_col = 0)
ola2_post_umi = pd.read_csv(path+"GSM6938170_ReSisTrace_Olaparib_2_post_treatment_UMIcounts.tsv.gz", sep='\t', index_col=0)

Find MTRNA genes

mtgenes = ["ENSG00000210049.1","ENSG00000211459.2","ENSG00000210077.1","ENSG00000210082.2","ENSG00000209082.1","ENSG00000198888.2","ENSG00000210100.1","ENSG00000210107.1","ENSG00000210112.1","ENSG00000198763.3","ENSG00000210117.1","ENSG00000210127.1","ENSG00000210135.1","ENSG00000210140.1","ENSG00000210144.1","ENSG00000198804.2","ENSG00000210151.2","ENSG00000210154.1","ENSG00000198712.1","ENSG00000210156.1","ENSG00000228253.1","ENSG00000198899.2","ENSG00000198938.2","ENSG00000210164.1","ENSG00000198840.2","ENSG00000210174.1","ENSG00000212907.2","ENSG00000198886.2","ENSG00000210176.1","ENSG00000210184.1","ENSG00000210191.1","ENSG00000198786.2","ENSG00000198695.2","ENSG00000210194.1","ENSG00000198727.2","ENSG00000210195.2","ENSG00000210196.2"]

Carboplatin pre-treatment

carbo1_pre_pctMT = carbo1_pre_umi.loc[mtgenes,:].sum(axis=0) / carbo1_pre_umi.sum(axis=0)*100
carbo1_pre_pctMT.name = "pctMT"
carbo1_pre = pd.concat([carbo1_pre_pctMT, carbo1_pre_info.loc[carbo1_pre_umi.columns,"drugSens"]], axis=1)

carbo2_pre_pctMT = carbo2_pre_umi.loc[mtgenes,:].sum(axis=0) / carbo2_pre_umi.sum(axis=0)*100
carbo2_pre_pctMT.name = "pctMT"
carbo2_pre = pd.concat([carbo2_pre_pctMT, carbo2_pre_info.loc[carbo2_pre_umi.columns,"drugSens"]], axis=1)

Carboplatin post-treatment

carbo1_post_pctMT = carbo1_post_umi.loc[mtgenes,:].sum(axis=0) / carbo1_post_umi.sum(axis=0)*100
carbo1_post_pctMT.name = "pctMT"
carbo1_post = pd.concat([carbo1_post_pctMT, carbo1_post_info.loc[carbo1_post_umi.columns,"drugSens"]], axis=1)

carbo2_post_pctMT = carbo2_post_umi.loc[mtgenes,:].sum(axis=0) / carbo2_post_umi.sum(axis=0)*100
carbo2_post_pctMT.name = "pctMT"
carbo2_post = pd.concat([carbo2_post_pctMT, carbo2_post_info.loc[carbo2_post_umi.columns,"drugSens"]], axis=1)

Merge replicates

carbo1_pre["status"] = "pre-treatment"
carbo1_post["status"] = "post-treatment"

carbo1 = pd.concat([carbo1_pre, carbo1_post], axis=0)
carbo1 = carbo1[carbo1.drugSens!="non"]
carbo1 = carbo1.dropna()

carbo2_pre["status"] = "pre-treatment"
carbo2_post["status"] = "post-treatment"

carbo2 = pd.concat([carbo2_pre, carbo2_post], axis=0)
carbo2 = carbo2[carbo2.drugSens!="non"]
carbo2 = carbo2.dropna()

carbo = pd.concat([carbo1, carbo2], axis=0)

pairs = [("pre-sensitive", "pre-resistant"), ("pre-resistant", "resistant")]

fig, ax = plt.subplots(1,1,figsize=(5,5))

ax = sns.violinplot(data=carbo, x="drugSens", y="pctMT", dodge=True,
order = ["pre-sensitive", "pre-resistant", "resistant"])
plt.axhline(y=carbo[carbo.drugSens == "pre-sensitive"].pctMT.median(), color='k', linestyle='--', linewidth=1.5)
                                        
annotator = Annotator(ax, pairs, data=carbo, x="drugSens", y="pctMT", 
                    order=["pre-sensitive", "pre-resistant", "resistant"])

annotator.configure(test='Mann-Whitney', text_format='star', text_offset=1, show_test_name=False)
annotator.apply_and_annotate()

fig.show()
fig.savefig(path  + "carboplatin.pdf", format="pdf")


Olaparib pre-treatment

ola1_pre_pctMT = ola1_pre_umi.loc[mtgenes,:].sum(axis=0) / ola1_pre_umi.sum(axis=0)*100
ola1_pre_pctMT.name = "pctMT"
ola1_pre = pd.concat([ola1_pre_pctMT, ola1_pre_info.loc[ola1_pre_umi.columns,"drugSens"]], axis=1)

ola2_pre_pctMT = ola2_pre_umi.loc[mtgenes,:].sum(axis=0) / ola2_pre_umi.sum(axis=0)*100
ola2_pre_pctMT.name = "pctMT"
ola2_pre = pd.concat([ola2_pre_pctMT, ola2_pre_info.loc[ola2_pre_umi.columns,"drugSens"]], axis=1)

Olaparib post-treatment

ola1_post_pctMT = ola1_post_umi.loc[mtgenes,:].sum(axis=0) / ola1_post_umi.sum(axis=0)*100
ola1_post_pctMT.name = "pctMT"
ola1_post = pd.concat([ola1_post_pctMT, ola1_post_info.loc[ola1_post_umi.columns,"drugSens"]], axis=1)

ola2_post_pctMT = ola2_post_umi.loc[mtgenes,:].sum(axis=0) / ola2_post_umi.sum(axis=0)*100
ola2_post_pctMT.name = "pctMT"
ola2_post = pd.concat([ola2_post_pctMT, ola2_post_info.loc[ola2_post_umi.columns,"drugSens"]], axis=1)

Merge replicates

ola1_pre["status"] = "pre-treatment"
ola1_post["status"] = "post-treatment"

ola2_pre["status"] = "pre-treatment"
ola2_post["status"] = "post-treatment"


ola1 = pd.concat([ola1_pre, ola1_post], axis=0)
ola1 = ola1[ola1.drugSens!="non"]
ola1 = ola1.dropna()

ola2 = pd.concat([ola2_pre, ola2_post], axis=0)
ola2 = ola2[ola2.drugSens!="non"]
ola2 = ola2.dropna()

olaparib = pd.concat([ola1, ola2], axis=0)

fig, ax = plt.subplots(1,1,figsize=(5,5))
plt.axhline(y=olaparib[olaparib.drugSens == "pre-sensitive"].pctMT.median(), color='k', linestyle='--', linewidth=1.5)

ax = sns.violinplot(data=olaparib, x="drugSens", y="pctMT", dodge=True,
order = ["pre-sensitive", "pre-resistant", "resistant"])
                                          
annotator = Annotator(ax, pairs, data=olaparib, x="drugSens", y="pctMT", 
order=["pre-sensitive", "pre-resistant", "resistant"])

annotator.configure(test='Mann-Whitney', text_format='star', text_offset=1, show_test_name=False)
annotator.apply_and_annotate()

fig.show()
fig.savefig(path  + "olaparib.pdf", format="pdf")


# MDAMB468 GSE228382

path = "/add/path/here/GSE228382_lineage/"

metadata = pd.read_csv(path + "Afa.TS.Seurat_metadata.txt", sep='\t', index_col = 0)

mtx = sc.read_mtx(path + "GSE228154_TS_matrix.mtx")
mtx = mtx.T

obs = pd.read_csv(path + "GSE228154_TS_barcodes.tsv", header=None)
obs.columns = ["cell_name"]
mtx.obs = obs

var = pd.read_csv(path + "GSE228154_TS_features.tsv", header=None, sep=' ')
var.columns = ["GeneID", "GeneName", "Type"]

mtx.var.index = var.GeneID
mtx.obs_names

Calculate pctMT

allG =  var.dropna()
mtgenes = [s for s in allG.GeneName.tolist() if s.startswith("MT-")]
mtgenes_id = allG[allG.GeneName.isin(mtgenes)].GeneID.tolist()

pctMT = mtx[:,mtx.var_names.isin(mtgenes_id)].X.sum(axis=1) / mtx.X.sum(axis=1)*100
mtx.obs["pctMT"] = pctMT.A1

Select sensitive and tolerant clones

mtx.obs.index = mtx.obs.cell_name

table3 = pd.read_csv(path + "Table_S3.txt", sep='\t', index_col=0)

tolerant = table3[table3["Afatinib  response  group"]=="Tolerant"].index.tolist()
sensitive = table3[table3["Afatinib  response  group"]=="Sensitive"].index.tolist()

status = ["tolerant" if x in tolerant else "sensitive" for x in metadata.linageBC]
metadata["status"] = status

mtx.obs = pd.concat([mtx.obs, metadata], axis=1)

Use treatment-naive cells (ctrl)

ctrl = mtx[mtx.obs.condition == "CTRL"]

Compare pctMT between sensitive, tolerant and the two top tolerant clones, described in the original paper

tol_clones = ["bc14-013:bc30-092942", "bc14-065:bc30-046216"]
ctrl.obs.loc[ctrl.obs.linageBC.isin(tol_clones),"status"] = "top_tolerant"

fig, ax = plt.subplots(1,1,figsize=(5,5))
ax = sns.violinplot(data=ctrl.obs, x="status", y="pctMT", dodge=True)
pairs = [("sensitive", "tolerant"), ("sensitive", "top_tolerant")]
annotator = Annotator(ax, pairs, data=ctrl.obs, x="status", y="pctMT", 
order = ["sensitive", "tolerant", "top_tolerant"])

annotator.configure(test='Mann-Whitney', text_format='star', text_offset=1, show_test_name=False)
annotator.apply_and_annotate()
plt.axhline(y=ctrl.obs[ctrl.obs.status == "sensitive"].pctMT.median(), color='k', linestyle='--', linewidth=1.5)

fig.show()
fig.savefig(path  + "afatinib.pdf", format="pdf")
